Move helper/keyframe-helper to engine/KeyframeCalcurator
