+++ packages/delir-core/src/engine/Task/ClipRenderTask.ts
@@ -1,6 +1,5 @@
 import * as _ from 'lodash'
 
-import * as KeyframeHelper from '../../helper/keyframe-helper'
 import { TypeDescriptor } from '../../plugin-support/type-descriptor'
 import { Clip } from '../../project'
 import { AssetPointerScheme } from '../../project/scheme/keyframe'
@@ -9,6 +8,7 @@ import { RealParameterValues, RealParameterValueTypes } from '../Engine'
 import { compileTypeScript } from '../ExpressionSupport/ExpressionCompiler'
 import * as ExpressionContext from '../ExpressionSupport/ExpressionContext'
 import ExpressionVM from '../ExpressionSupport/ExpressionVM'
+import * as KeyframeHelper from '../KeyframeCalcurator'
 import * as RendererFactory from '../Renderer'
 import { IRenderer } from '../Renderer/RendererBase'
 import RenderRequest from '../RenderRequest'
